changed all uint colors to int colors possibly fixing any weird conversation issues

+++ Assembly-CSharp/Custom/Wardrobe/CE_WardrobeManager.cs
@@ -896,7 +896,7 @@ public class CE_WardrobeManager
 			if (useShader)
             {
 				spriteRenderer.material = HatShaderMaterial;
-				PlayerControl.SetPlayerMaterialColors((uint)colorId, spriteRenderer);
+				PlayerControl.SetPlayerMaterialColors(colorId, spriteRenderer);
 			}
 			else
             {
